refactor/ArticleActions: ArticleActions has been moved to a separate folder
