+++ src/components/Article/ArticleActions/ArticleActions.js
@@ -1,8 +1,8 @@
 import { Link } from 'react-router-dom';
 import React from 'react';
-import agent from '../../agent';
+import agent from '../../../agent';
 import { connect } from 'react-redux';
-import { DELETE_ARTICLE } from '../../constants/actionTypes';
+import { DELETE_ARTICLE } from '../../../constants/actionTypes';
 
 const mapDispatchToProps = dispatch => ({
   onClickDelete: payload =>
